requirements.txt
flask>=2.0
Pillow>=9.0
numpy>=1.21
opencv-python>=4.5